Bug #1347: quota update accepts float values (cherry picked from commit c94bfcab5e819b0f494a4f7c31ad3b7dd8ed1a22)

+++ include/Quota.h
@@ -119,7 +119,7 @@ protected:
      *    @param usage_req usage for each metric
      */
     void del_quota(const string& qid, 
-                   map<string, int>& usage_req);
+                   map<string, float>& usage_req);
 
     /**
      *  Gets a quota identified by its ID.
